:recycle:関数・ファイル分割
